Download a file from Object Storage

Download a file from Object Storage

dwnldFromObjSto(prefix, bucket, path)

Arguments

prefix

passed to aws.s3::s3sync(). Must be a path to a remote folder. Use trailing slash (~/).

bucket

passed to aws.s3::s3sync().

path

passed to aws.s3::s3sync(). Must be an existing path to a local folder.

Details

It is assumed that th object storage connection details and credentials have already been set up as system environment variables (notably "AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY" and "AWS_S3_ENDPOINT"). Note that all objects inside path will be synced to path